+++ src/object-store/Cargo.toml
@@ -12,8 +12,12 @@ services-memory = ["opendal/services-memory"]
 
 [dependencies]
 bytes.workspace = true
+common-base.workspace = true
+common-error.workspace = true
+common-macro.workspace = true
 common-telemetry.workspace = true
 futures.workspace = true
+humantime-serde.workspace = true
 lazy_static.workspace = true
 md5 = "0.7"
 moka = { workspace = true, features = ["future"] }
@@ -28,6 +32,9 @@ opendal = { version = "0.52", features = [
     "services-s3",
 ] }
 prometheus.workspace = true
+reqwest.workspace = true
+serde.workspace = true
+snafu.workspace = true
 tokio.workspace = true
 uuid.workspace = true
